Add the concept of disabled categories
Each runner can provide a list of categories which can be used to
provide the runner extra information about which categories should be
given. This is quite useful when a single runner provides many different
categories and not all are required.

Additionally, the RunnerManager also disables runners whose categories
are disabled.

REVIEW: 118148