drop examples dir
Summary:
- it does not build on its own (doesn't find_package anything)
- it is not wired into the overall build (so you can't build it as part of
  that either)
- it does not get installed (making it nigh impossible to find for 3rd
  party devs)
- 2/2 kde devs did not even know this example exists
- it contains a Messages.sh which is run and extracts and even gets
  translations (! what waste !)

to fix all of the above drop the entire directory. if someone wants to
bring it back, the above issue list should be dealt with. in particular
the Messages.sh problem (conceivably the extract line could be commented
out with a todo note that it needs enabling for i18n).

Reviewers: broulik

Reviewed By: broulik

Subscribers: ltoscano, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5032