Add a done signal to FindMatchesJob instead of using QObjectDecorator wrongly
Summary:
Currently KRunner uses QObjectDecorator wrongly and changing the
design to fix that would not only be a lot of work, but also include downsides
like having two custom Job classes, making the code harder to read.
So merge the part of QObjectDecorator that would've been used here into the
FindMatchesJob class and drop use of QObjectDecorator.

Test Plan: Added a debug statement in the jobDone slot, it's called.

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22758