DBus Runner: Add service property to request actions once
Summary:
With the `X-Plasma-Request-Actions-Once` property the plugin
can specify if the actions should only be requested when the plugin is initialized.

This is useful if the plugin always used the same actions, because
then the we don't need to make the dbus calls for every match session.

Test Plan:
Create dbus runner and don't set the extra property, the actions should be requested for each session.
(More like each query because of BUG: 420311).
After setting the property to true the actions should only be requested when the runner is initialized.

Reviewers: #plasma, meven, ngraham, broulik

Reviewed By: meven, ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29175