KRunner fix prepare/teardown signals
Summary:
BUG: 420311.
Because the teardown was requested and checked the prepped variable was set to false and consequently
the prepare method gets called for the next character typed.

The exact two lines get called in the `matchSessionComplete` method.
The `matchSessionComplete` method gets called called in `RunnerResultsModel::clear` (milou repo).

Test Plan:
If you connect to the prepare/teardown slot, for example:
```
connect(this, &MyRunner::prepare, [=]() {
        qWarning() << "prepare";
});````

You see that they get triggered for each letter typed, after applying this patch they get only triggered when the match session is started/is over.

Reviewers: meven, ngraham, broulik

Reviewed By: meven

Subscribers: davidedmundson, cfeck, kde-frameworks-devel

Tags: #frameworks, #plasma

Differential Revision: https://phabricator.kde.org/D29050