Simplify SingleRunnerMode code
The code has the concept that if there is a query in the single runner
mode and there are are no runners loaded the single runner
will be loaded and deleted after use.

This does not match our practical usecase at all and just
makes the codebase more complex.

And it would also be slower, because the same runner needs to
be loaded again, when all the runners are loaded.

To test this change out you can run:
`qdbus org.kde.krunner /App org.kde.krunner.App.querySingleRunner bookmarks kde`
and you should only get results from the bookmarks runner.