Do not change relevance based on launch counts
This cases big problems, because the relevance that the runner
sets gets modified. Practically this means that matches get
a high relevance, even if these do not match the search term
at all.

Considering that this change goes in right before tagging
I want to keep the change as small as possible and will do the
cleanup later.

BUG: 432339
BUG: 431609

FIXED-IN: 5.79