Allow runners to prevent duplicate results
This commit allows the runners to use the id to prevent duplicate
results. For this the X-Plasma-Runner-Unique-Results property is added.
If the runners set this to true the runner name is not prepended to the id and
it is checked if a match with the id already exists before the match is added.

Because the property introduces a behavior change we can not enable it
by default.

In case there are duplicate results the results from the runner which has the X-Plasma-Runner-Weak-Results property
set to true will be removed. Otherwise the results from the first runner
will be used.

BUG: 406292
FIXED-IN: Frameworks 5.82 & Plasma 5.22