dbusrunner: Add Teardown and Config methods to D-Bus interface
This adds the prepare and teardown lifecycle
methods to the interface. This aligns with the
goal of making DBus runners first class citizens.

The Config method can be used to set config properties and in the future
additional properties at runtime.

Also the X-Plasma-Request-Actions-Once property is evaluated in a per-service bases
and not a per-plugin one. This makes it work more reliable with wildcard services.

Bug: 435350
FIXED-IN: 5.85