dbusrunner: Make X-Plasma-Request-Actions-Once work more reliable
In case we have wildcard services that are not running when KRunner
is started the actions would not be fetched. Now the actions are
fetched at the beginning of a match session. To make sure
the X-Plasma-Request-Actions-Once property is respected a set with the
already fetched service names is used.

BUG: 435350
FIXED-IN: 5.84