runnermanager: Remove logic for deferred runs
This logic came into play when a runner's match is run, but the
runner has not finished matching. In this case the match would be run
after the runner has finished matching.

This could make sense if the runner has to set some internal state
first, but the matches can contain their own data and the
AbstractRunner::match method is not thread safe anyways. Also the method
docs strictly advise against using a state which gets modified in the
match method.