Use KFileUtils::findAllUniqueFiles to locate DBus runners
This takes care of de-duplicating the files and prefers the most local one.

Also we don't need the std::function callback anymore and can
just use a simple for loop. This will also be helpful when we can
drop the runnerId de-duplication checks after the old KServiceTypeTrader
code path can be dropped.