Allow runners to opt-out storage of entry to history
In case of the help runner, the queries are automatically completed.
Consequently there is no point in showing the suggestions for it.

Also it could make sense to opt-out the history for runners that
use privacy sensitive info. For example the webshortcuts runner
opening an URL in a private window.