Fix flickering in Application Launcher for every character typed
This fixes the flickering that happened on the Application Launcher for every character typed.

Before:
![before](/uploads/cfd6dff218256a4de398deb9982e40df/before.gif)
After:
![after](/uploads/34a6aff7cc79dcb51f102c07573353ae/after.gif)

`RunnerManagerPrivate::scheduleMatchesChanged()` method forwards its events only once every 250ms. The issue was that it "wasted" the very first refresh cycle by forwarding the initially empty matches result set. It only forwarded meaningful results after 250ms, that's why the Application Launcher showed an empty list for 250ms for every character typed.

We are now stalling for the duration of the timeout (250ms) before starting to display results for a new query, so we kind of "ignore" the empty/near-empty result set (unless it actually hits the timeout with no more results coming in the meantime). This avoids refreshing the client with the initially empty (or near-empty) result set, so users won't see the empty list if more meaningful results come in the next 250ms (and they usually do come!).

I also saw that when the search job was done, the engine still waited for the running timeout until it displayed the last results it found. I changed that as well, so now we intercept the running timer, cancel it, and refresh immediately as soon as the job is done. So if the total query runs in less than 250ms, it won't even have to wait that: it will refresh only once (as intended), with the full result set, and as fast as possible.

As a result of all this, the application launcher feels much more responsive now, I believe it is a meaningful improvement in user experience.

Also benefits the standalone krunner since it will now respond faster than 250ms for simple quick queries.

BUG: 423161