DBusRunner: Use watcher as context for DBus match connection
Otherwise we keep the function & captured variables around until the runner is unloaded.
This is obviously not what we want, instead we want to delete the function once the matching is finished.

Because we wait for all requests to finish, we can be sure that we do not have edge conditions where the finished
signal is emitted while the object is destroyed.