Remove upper-limits for RunnerManagerTest's timeouts
Those upper-limits timeouts were not conservative enough for the Continuous Integration environment which can be slow and this test was failing sometimes.

The upper-limits timing verifications were removed without too much prejudice to the overall test. It should still serve its purpose to guarantee there will be no regression on this functionality.