Make Qt::Gui dependency for deprecation free builds internal
We include this in the AbstractRunner file, but fwd it in the QueryMatch file.
However one can perfectly work without haveing this dependency public, because
the icon name can be set directly. Runners usually do not need to access the QIcon of the runner
itself, because that is implicitly set when one does not define an icon for the query match.