Add support for XDG actiovation to DBus Runners
Executing a runner action may involve rasing an existing window, for example when activating a room result from the NeoChat runner.

To be able to do this on Wayland we need to request an activation token and pass that to the runner.
The runner then can use that to raise its window