Import RunnerResultsModel model from milou
The model is quite an important part of the KRunner ecosystem and thus it makes sense
to have it in the framework. For now, it is QML-only, but we can export it when needed.
A possible usecase is the model that kicker uses.

This should also make potential refactorings for the sorting easier, because all components except the plugins
are in one central place.