Make QueryMatch::setSelected action internal, take action in RunnerManager::run
Theoretically, a runner could set the selected action when creating a match.
This makes absolutely no sense, because the selected action must be set according to
the user's interaction.
Also, a QueryMatch resembles a possibility the user can run and is not intended to contain
information about the action the user took once the matches were displayed.

Because adding a QAction* arg in AbstractRunner::run could not be doable in a
source code compatible way, only the setter in QueryMatch is made internal.
Users of the KRunner framework can use the added action arg in RunnerManager::run
to provide the selected action by the user.