Simplify running of matches
Previously:
In the actionsRepeater from milou we react to click events on the actions.
In the event handler we tell the model to run this action.
The model sets the selected action and calls RunnerManager::run.
The runner context does some stuff and delegates the running to the RunnerContext. We do this, because the RunnerContext remembers the often launches matches.
And finally the RunnerContext calls QueryMatch::run which calls the runner's run impl

Now:
The model/milou bits are the same.
However, we now call AbstractRunner::run directly in RunnerManager::run
and tell the RunnerContext afterwards that we have run a match.