Suspend runners while initializing/reloading config
The initialization may take longer, in that time, there is no point in
launching queries.
Also, this avoids thread-safety issues because the RunnerManager reads
the minlettercount/matchregex from the AbstractRunner.

Because we suspend runners more often now, we must not emit
queryFinished while runners are suspended. Otherwise, we emit
queryFinished and emit lots of results later on.
This might be visually confusing and makes testing difficult and error
prone.