RunnerManager: Add X-KRunner-Private-MainThread property to not move runner to own thread
For example the places runner can't be moved to their own thread.
In case runners are fully async, they could use this property too and just use a nested eventloop as before.
Or we could utilize this in some trivial runners that are not worth to have their own thread (the plasma-desktop runners)