Correct invalid DBus type signature annotation in Config method.
Latest Qt 6 has updated qtdbuscpp2xml to fail and report errors when the
DBus XML has invalid type signature or invalid DBus path, object or
method names, which will cause the build to fail if it occurs.

The stricter checking detected an error in the org.kde.krunner1.Config
method's result type annotation. Per the D-Bus specification, a bare
DICT_ENTRY signature (the {..} with appropriate key and value types) is
invalid, it needs to be part of an ARRAY of DICT_ENTRY to be valid.

Adjusting the type to be an ARRAY of DICT_ENTRY fixes the error and is
compatible with the actual output type (QVariantMap).

This fixes the build, however I'm not currently able to test because I'm
still working to get my KF6 Plasma setup.