RunnerContext: Do not save context after every match session
We don't run a match in every session, meaning if we close KRunner
without running any match, we'd write it anyway.

Also, writing it for every run is not necessary. The idea is to not rely
on the syncing on app exit, because that caused the config to not be
written in some cases. As a middle ground, the value is synced every
5 runs.
For tests, the variable has been exported so that it can be changed to
sync the config immediately.