QueryMatch: Avoid usage of recursive QReadWriteLock
Just like with a recursive mutex, this comes with a performance cost.
In our case, we only needed it for one function. This can be easily
moved to the d-ptr and the locking code is in the public class.

Also, don't create a pointer that we have to clean up manually, instead,
define it as a normal member variable and get its address when
constructing the lock.