Allow to set RunnerManager instance in model from outside
This allows for reusing the manager instances or keeping them separate from the lifecycle of the model.
A usecase if KWin, where we want to have the Milou.ResultView on
different screens, but still want to keep the query and thus datasource
in sync.
Otherwise, we would do the exact same queries twice and in case of the
calculator get slight issues with the global mutex.

CCBUG: 483147