Revert "Give favorites a relative boost rather than absolute sorting position"
This reverts commit d02e1c13a9ee92a90218a4584a3642f39da4e28a. It
undermines the manual sorting feature by causing search result
ordering to not conform to the user's configured preference.

If the user wants results ordered in a particular way, KRunner should
always honor that.

If this leads to an unacceptable UX, then it means the manual ordering
feature itself is unworkable and should be removed instead.

BUG: 489866
FIXED-IN: 6.13