manager,context: remove launchcount adjustment
while theoretically a good idea on paper, this makes for super awkward
to debug behavior between applications **and** users/systems.
when searching for dolphin 'do' in kickoff it could rank ksudoku first
but when doing the same in krunner it could rank dolphin first, entirely
dependent on a magic number that nobody knows about. this gets even more
fun when you have different users compare results, such as in a bug
report.

there are certainly ways to implement this feature in a smarter fashion
(global, contextually bound to a query, etc), but ultimately the problem
of reproducibility remains

let's remove this instead. it was a very niche feature that I am sure
had its fans, but on the balance, muscle memory probably achieves the
same end result